Corrección de intersecciones y desvío de carros

+++ Simulacion/src/movimiento/MovimientoUniforme.scala
@@ -4,7 +4,7 @@ trait MovimientoUniforme extends Movil {
   def aumentarPosicion(dt:Double):Unit = {   
     val velx = math.cos(velocidad.direccion.grados.toRadians).toDouble * velc.toDouble
     val vely = math.sin(velocidad.direccion.grados.toRadians).toDouble * velc.toDouble
-    this.posicion.x_((posicion.x.toDouble + velx * dt).toInt)
-    this.posicion.y_((posicion.y.toDouble + vely * dt).toInt)
+    this.posicion.x_(posicion.x.toDouble + velx * dt)
+    this.posicion.y_(posicion.y.toDouble + vely * dt)
   }  
 }
